Replace i32 with u32 and change `try_into` to `as`

+++ src/i3.rs
@@ -51,7 +51,7 @@ impl Core<I3Connection, I3Config> {
     fn handle_window_event(&mut self, event_info: WindowEventInfo) {
         let node = event_info.container;
         let id = match node.window {
-            Some(x) => x,
+            Some(x) => x as u32,
 
             // It means, the window was sent to scratchpad desktop
             None => {
